+++ R/zzz.R
@@ -1,5 +1,5 @@
 #' Internal
 #' @noRd 
 #' @importFrom utils globalVariables
-globalVariables(c("time", "times", "value", "variable"))
+globalVariables(c("leading_ev", "time", "times", "value", "variable"))
 globalVariables(c("a_12", "a_21", "res_stab"))