+++ tests/testthat/test_infer.R
@@ -49,12 +49,18 @@ test_that("fw_infer() format errors", {
 })
 
 
-test_that("", {
+test_that("fw_infer() works", {
     suppressWarnings({
         res1 <- fw_infer(A, R, B)
         res2 <- fw_infer(A, R, B, iter = 1000)
     })
     expect_true(inherits(res1, "data.frame"))
+    expect_true(inherits(res1, "fw_predicted_int"))
+    expect_true(all(res1$leading_ev < 0))
     expect_identical(dim(res1), c(3000L, 4L))
     expect_identical(dim(res2), c(1000L, 4L))
+    expect_error(fw_get_B_predicted(data.frame(1), A, R))
+    expect_equal(fw_get_B_predicted(res1[1, ], A, R), B)
+    res3 <- fw_get_A_predicted(res1[1, ], A)
+    expect_identical((res3 > 0) * 1.0 - (res3 < 0) * 1.0, A)
 })